requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pandas>=2.0
numpy>=1.24
faiss-cpu>=1.7.4
sentence-transformers>=3.0
rank-bm25>=0.2.2
scikit-learn>=1.3
pydantic>=2.0
httpx>=0.24
